Add admin role pages and update routing; refactor sidebar menu for role-based navigation
