+++ src/pages/admin-role/contests/Contests.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Contests = () => {
+    return (
+        <div>Contests</div>
+    )
+}
+
+export default Contests
