+++ src/routes/PrivateRoute.js
@@ -18,7 +18,7 @@ const PrivateRoute = ({ allowedRoles }) => {
   if (role?.toLowerCase() === 'user') {
     return <Navigate to="/participant/home" replace />;
   } else if (role?.toLowerCase() === 'admin') {
-    return <Navigate to="/admin/dashboard" replace />;
+    return <Navigate to="/admin/home" replace />;
   }
 
   return <Navigate to="/sign-in" />;
